+++ ui/src/GraphPage.vue
@@ -1,13 +1,15 @@
 <script setup lang="ts">
 import { computed } from 'vue';
 import { useApp } from './app';
-import { GraphMakerSettings, GraphMaker } from '@milaboratories/graph-maker';
+import { GraphMakerState, GraphMaker } from '@milaboratories/graph-maker';
+import { GraphPageState } from '@platforma-open/milaboratories.graph-maker.model';
 
 const app = useApp<`/graph?id=${string}`>();
 
 const ui = app.createUiModel(undefined, () => ({ graphs: [] }));
 
-const settings = computed({
+const graphProps = computed(() => ui.model.graphs.find(it => it.id === app.queryParams.id)?.settings)
+const state = computed({
   get() {
     const graphState = ui.model.graphs.find(it => it.id === app.queryParams.id);
     if (!graphState) {
@@ -15,18 +17,18 @@ const settings = computed({
       return null;
     }
     return {
-      ...(graphState.settings as GraphMakerSettings),
-      allowDeleting: true,
+      ...(graphState.state as GraphMakerState),
       title: graphState.label
-    } as GraphMakerSettings;
+    } as GraphMakerState;
   },
-  set(nextSettings: GraphMakerSettings) {
+  set(nextState: GraphMakerState) {
     ui.model.graphs = ui.model.graphs.map((item) => {
       return item.id === app.queryParams.id ? {
         id: item.id,
-        settings: nextSettings,
-        label: nextSettings.title
-      } : item;
+        label: nextState.title,
+        state: nextState,
+        settings: graphProps.value,
+      } as GraphPageState : item;
     });
   }
 });
@@ -44,15 +46,21 @@ const removeSection = async () => {
     app.navigateTo('/');
   }
 };
-
+//
+// watch(() => app.model.outputs.pFrame, async (handle) => {
+//   const list = await platforma?.pFrameDriver.listColumns(handle!);
+//   console.log(list, 'list')
+// }, {immediate: true})
 </script>
 
 <template>
   <div class="container_graph_page" :key="app.queryParams.id">
     <graph-maker
-      v-if="settings"
-      v-model="settings"
+      v-if="state && graphProps"
+      v-model="state"
       :pFrame="app.model.outputs.pFrame"
+      :chartType="graphProps.chartType"
+      :allowDeleting="true"
       @delete-this-graph="removeSection"
     />
   </div>