+++ block/CHANGELOG.md
@@ -1,5 +1,137 @@
 # @platforma-open/milaboratories.graph-maker
 
+## 1.4.11
+
+### Patch Changes
+
+- Updated dependencies [470e969]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.9
+
+## 1.4.10
+
+### Patch Changes
+
+- Updated dependencies [9c67b2a]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.8
+
+## 1.4.9
+
+### Patch Changes
+
+- Updated dependencies [877dc37]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.7
+
+## 1.4.8
+
+### Patch Changes
+
+- Updated dependencies [7abbfa6]
+- Updated dependencies [7abbfa6]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.6
+
+## 1.4.7
+
+### Patch Changes
+
+- Updated dependencies [4bb138d]
+  - @platforma-open/milaboratories.graph-maker.model@1.2.3
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.5
+
+## 1.4.6
+
+### Patch Changes
+
+- Updated dependencies [21ea27d]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.4
+
+## 1.4.5
+
+### Patch Changes
+
+- 85ba82f: chore: fix changelog
+
+## 1.4.4
+
+### Patch Changes
+
+- Updated dependencies [59abd3d]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.3
+
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies [77056ac]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.2
+
+## 1.4.2
+
+### Patch Changes
+
+- 419f284: chore: update marketplace score
+
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [eef8887]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.1
+
+## 1.4.0
+
+### Minor Changes
+
+- 27911fa: chore: bump version
+
+## 1.3.0
+
+### Minor Changes
+
+- 4ffcddf: chore: update ci wf
+
+### Patch Changes
+
+- Updated dependencies [4ffcddf]
+  - @platforma-open/milaboratories.graph-maker.ui@0.3.0
+
+## 1.2.18
+
+### Patch Changes
+
+- Updated dependencies [e689add]
+- Updated dependencies [6448ec4]
+  - @platforma-open/milaboratories.graph-maker.ui@0.2.15
+
+## 1.2.17
+
+### Patch Changes
+
+- Updated dependencies [a9c8e94]
+  - @platforma-open/milaboratories.graph-maker.ui@0.2.14
+
+## 1.2.16
+
+### Patch Changes
+
+- Updated dependencies [b753e49]
+  - @platforma-open/milaboratories.graph-maker.ui@0.2.13
+
+## 1.2.15
+
+### Patch Changes
+
+- Updated dependencies [24f6b2e]
+  - @platforma-open/milaboratories.graph-maker.ui@0.2.12
+
+## 1.2.14
+
+### Patch Changes
+
+- Updated dependencies [db01bb7]
+- Updated dependencies [54ba316]
+  - @platforma-open/milaboratories.graph-maker.model@1.2.2
+  - @platforma-open/milaboratories.graph-maker.ui@0.2.11
+
 ## 1.2.13
 
 ### Patch Changes