+++ model/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @platforma-open/milaboratories.graph-maker.model
 
+## 1.2.3
+
+### Patch Changes
+
+- 4bb138d: add heatmap + dendro template
+
+## 1.2.2
+
+### Patch Changes
+
+- db01bb7: update dependencies
+- 54ba316: update graph-maker component interface
+
 ## 1.2.1
 
 ### Patch Changes